src/validator/requirements.txt
pandas==2.2.2
pyyaml==6.0.2
rich==13.7.1
